feat(v11): 开放浮窗并接入布局 V3 自动保存与安全退出
启用单项和整组浮动，保留主骨架、跨窗口命令、范围关闭和原实例迁移。生产切换严格 V3，增加串行防抖写入、最终快照冻结、只读锁、找回与重置回滚；Runtime 在释放工作区前停止布局调度。补原生窗口、边界、竞态及跨进程验证，同步新结构下的专用指南和契约。完整开发门禁在文档定稿后执行，真机与发布状态独立保留。

+++ Host/MyAvaloniaManagement/Business/Layout/DockLayoutV3Json.cs
@@ -42,7 +42,7 @@ internal static class DockLayoutV3Json
         using var json = JsonDocument.Parse(buffer.ToArray(), new JsonDocumentOptions { MaxDepth = 100 });
         CheckDuplicates(json.RootElement);
         if (json.RootElement.ValueKind == JsonValueKind.Object && json.RootElement.TryGetProperty("schemaVersion", out var version) &&
-            version.ValueKind == JsonValueKind.Number && version.TryGetInt32(out var schema) && schema > 3)
+            version.ValueKind == JsonValueKind.Number && version.TryGetDouble(out var schema) && schema > 3)
             throw new DockLayoutFormatException("LAYOUT_SCHEMA_FUTURE");
         try
         {
